remove warning

+++ src/main.cpp
@@ -4,5 +4,5 @@
 int main() {
     auto s = std::string("Hello"); // checks for C++ std and C++ 11
     std::cout << s << std::endl;
-    return s.length();
+    return (int)s.length();
 }
